camera-multi-streaming/requirements.txt
farm-ng-amiga
opencv-python==4.6.0.66
numpy==1.26.4
